Bugfixes: -First column in Excel was interpreted as language -File.Delete was called when file didn't exist

+++ i18n-converter/JsonConverter.cs
@@ -89,7 +89,7 @@ namespace I18nConverter
                     }
                     
                     var filePath = Path.Combine(outPath, language, i18nNamespace + ".json");
-                    if (!languageDictionary.Any())
+                    if (!languageDictionary.Any() && File.Exists(filePath))
                     {
                         File.Delete(filePath);
                         continue;
